+++ src/gmp.rs
@@ -881,10 +881,20 @@ extern "C" {
 pub unsafe fn mpq_numref(op: mpq_ptr) -> mpz_ptr {
     (&mut (*op).num) as mpz_ptr
 }
+/// Constant version of [`mpq_numref`](#method.mpq_numref).
+#[inline]
+pub unsafe fn mpq_numref_const(op: mpq_srcptr) -> mpz_srcptr {
+    (&(*op).num) as mpz_srcptr
+}
 /// See: [`mpq_denref`](https://tspiteri.gitlab.io/gmp-mpfr/gmp/Rational-Number-Functions.html#index-mpq_005fdenref)
 #[inline]
 pub unsafe fn mpq_denref(op: mpq_ptr) -> mpz_ptr {
     (&mut (*op).den) as mpz_ptr
+}
+/// Constant version of [`mpq_denref`](#method.mpq_denref).
+#[inline]
+pub unsafe fn mpq_denref_const(op: mpq_srcptr) -> mpz_srcptr {
+    (&(*op).den) as mpz_srcptr
 }
 extern "C" {
     /// See: [`mpq_get_num`](https://tspiteri.gitlab.io/gmp-mpfr/gmp/Rational-Number-Functions.html#index-mpq_005fget_005fnum)